processfullvcf.sh currently runs on a vcf file and removes indels (I think. Not 100% sure there were indels in the test inputs), heterozygous sites, and any uncalled sites. Uncommenting lines will make it into a script for first merging a list of vcf files and then doing all that cleaning.
snp2seq.py was originally written by Murat. I cleaned it up according to the Google style sheet for Python to improve its readability, and I also changed a string append to a list append, because that is supposed to be much faster. The script now takes an input VCF file and produces a FASTA file of the sample sequences. It is suitable for processing a clean multi-sample VCF file into what is essentially a FASTA alignment.